monitor/requirements.txt
schedule==1.1.0
kafka-python==2.0.2